♻️ refactor : WebSocket 역할에 부합하게 기존 repository -> service로 수정
